# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.OpenAIError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  docs.OpenAIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/docs.OpenAIError'
    type: object
...
  docs.OpenAIModel:
    properties:
      created:
        example: 1626777600
        type: integer
      id:
        example: gpt-4o-mini
        type: string
      object:
        example: model
        type: string
      owned_by:
        example: openai
        type: string
      parent:
        example: ""
        type: string
      permission:
        items:
          $ref: '#/definitions/docs.OpenAIModelPermission'
        type: array
      root:
        example: gpt-4o-mini
        type: string
    type: object
...
  docs.OpenAIModelPermission:
    properties:
      allow_create_engine:
        example: true
        type: boolean
      allow_fine_tuning:
        example: false
        type: boolean
      allow_logprobs:
        example: true
        type: boolean
      allow_sampling:
        example: true
        type: boolean
      allow_search_indices:
        example: false
        type: boolean
      allow_view:
        example: true
        type: boolean
      created:
        example: 1626777600
        type: integer
      group:
        example: default
        type: string
      id:
        example: modelperm-abc123
        type: string
      is_blocking:
        example: false
        type: boolean
      object:
        example: model_permission
        type: string
      organization:
        example: '*'
        type: string
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
...
  /api/v1/public/assistants:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/assistants/{id}:
    delete:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/assistants/{id}/files:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/assistants/{id}/files/{fileId}:
    delete:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
...
  /api/v1/public/files:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/files/{id}:
    delete:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/files/{id}/content:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/fine_tuning/jobs:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/fine_tuning/jobs/{id}:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/fine_tuning/jobs/{id}/cancel:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/fine_tuning/jobs/{id}/events:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
...
  /api/v1/public/images/edits:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
...
  /api/v1/public/images/variations:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
...
  /api/v1/public/models/{model}:
    delete:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    get:
      parameters:
      - description: Model ID
        in: path
        name: model
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.OpenAIModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve model detail (OpenAI compatible)
      tags:
      - public
...
  /api/v1/public/threads:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}:
    delete:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/messages:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/messages/{messageId}:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/messages/{messageId}/files:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/messages/{messageId}/files/{filesId}:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/runs:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/runs/{runsId}:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/runs/{runsId}/cancel:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/runs/{runsId}/steps:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/runs/{runsId}/steps/{stepId}:
    get:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
  /api/v1/public/threads/{id}/runs/{runsId}/submit_tool_outputs:
    post:
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/docs.OpenAIErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible endpoint not implemented
      tags:
      - public
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"